requirements.txt
numpy==1.15.4
pandas==0.23.4
psycopg2-binary==2.7.6.1
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
